requirements.txt
amqp==2.2.2
appnope==0.1.0
awsebcli==3.12.1
billiard==3.5.0.3
blessed==1.14.2
boto3==1.4.7
botocore==1.7.45
celery==4.1.0
cement==2.8.2
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.7
coolsms-python-sdk==2.0.3
decorator==4.1.2
Django==1.11.6
django-appconf==1.0.2
django-cors-headers==2.1.0
django-extensions==1.9.1
django-imagekit==4.0.2
django-storages==1.6.5
djangorestframework==3.7.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
future==0.16.0
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
jmespath==0.9.3
kombu==4.1.0
olefile==0.44
parso==0.1.0
pathspec==0.5.0
pexpect==4.2.1
pickleshare==0.7.4
pilkit==2.0
Pillow==4.3.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.9.1
s3transfer==0.1.11
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
websocket-client==0.44.0
